cleanup: move code around
