+++ fedimint-server-ui/src/dashboard/bitcoin.rs
