+++ fedimint-server-ui/src/dashboard/general.rs
