+++ fedimint-server-ui/src/dashboard/latency.rs
