+++ fedimint-server-ui/src/dashboard/modules/meta.rs
@@ -1,6 +1,7 @@
 use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet};
 
-use axum::extract::{Form, State};
+use axum::extract::{Form, FromRequest, State};
+use axum::http::StatusCode;
 use axum::response::{Html, IntoResponse, Redirect, Response};
 use axum_extra::extract::cookie::CookieJar;
 use fedimint_core::PeerId;
@@ -8,9 +9,10 @@ use fedimint_core::module::serde_json::{self, Value};
 use fedimint_meta_server::Meta;
 use fedimint_server_core::dashboard_ui::{DashboardApiModuleExt, DynDashboardApi};
 use maud::{Markup, html};
-use tracing::warn;
+use serde::Serialize;
+use thiserror::Error;
+use tracing::{debug, warn};
 
-use crate::error::{RequestError, RequestResult};
 use crate::{AuthState, LOG_UI, check_auth};
 
 // Function to render the Meta module UI section
@@ -396,4 +398,53 @@ pub fn render_meta_edit_form(
             }
         }
     }
+}
+
+/// Wrapper over `T` to make it a json request response
+#[derive(FromRequest)]
+#[from_request(via(axum::Json), rejection(RequestError))]
+struct AppJson<T>(pub T);
+
+impl<T> IntoResponse for AppJson<T>
+where
+    axum::Json<T>: IntoResponse,
+{
+    fn into_response(self) -> Response {
+        axum::Json(self.0).into_response()
+    }
+}
+
+/// Whatever can go wrong with a request
+#[derive(Debug, Error)]
+pub enum RequestError {
+    #[error("Bad request: {source}")]
+    BadRequest { source: anyhow::Error },
+    #[error("Internal Error")]
+    InternalError,
+}
+
+pub type RequestResult<T> = std::result::Result<T, RequestError>;
+
+impl IntoResponse for RequestError {
+    fn into_response(self) -> Response {
+        debug!(target: LOG_UI, err=%self, "Request Error");
+
+        let (status_code, message) = match self {
+            Self::BadRequest { source } => {
+                (StatusCode::BAD_REQUEST, format!("Bad Request: {source}"))
+            }
+            _ => (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                "Internal Service Error".to_owned(),
+            ),
+        };
+
+        (status_code, AppJson(UserErrorResponse { message })).into_response()
+    }
+}
+
+// How we want user errors responses to be serialized
+#[derive(Serialize)]
+pub struct UserErrorResponse {
+    pub message: String,
 }
