+++ fedimint-server-ui/src/dashboard/modules/mod.rs
@@ -1,0 +1,3 @@
+pub mod lnv2;
+pub mod meta;
+pub mod wallet;
