+++ fedimint-server-ui/src/dashboard/modules/wallet.rs
