--- fedimint-server-ui/src/layout.rs
@@ -1,19 +1,0 @@
-use maud::{Markup, html};
-
-pub fn common_head(title: &str) -> Markup {
-    html! {
-        meta charset="utf-8";
-        meta name="viewport" content="width=device-width, initial-scale=1.0";
-        title { "Guardian Dashboard"}
-        link rel="stylesheet" href="/assets/bootstrap.min.css" integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous";
-        link rel="stylesheet" type="text/css" href="/assets/style.css";
-        link rel="icon" type="image/png" href="/assets/logo.png";
-
-        // Note: this needs to be included in the header, so that web-page does not
-        // get in a state where htmx is not yet loaded. `deref` helps with blocking the load.
-        // Learned the hard way. --dpc
-        script defer src="/assets/htmx.org-2.0.4.min.js" {}
-
-        title { (title) }
-    }
-}
